feat: fixed splint type mismatch error

+++ symtablehash.c
@@ -45,7 +45,7 @@ SymTable_T SymTable_new(void)
 {
     SymTable_T oSymTable;
     size_t i;
-    const int INITIALIZATION_BUCKETS = 509;
+    const size_t INITIALIZATION_BUCKETS = 509;
     struct SymTableNode **ppsFirstNode;
 
     oSymTable = (SymTable_T)malloc(sizeof(struct SymTable));
